input refactored
